1.1

+++ bot/handlers/commands.py
@@ -5,7 +5,7 @@ from bot.database.mysql import mysql
 from bot.database.sqlite import sqlite
 from bot.keyboards import inline
 from bot import states as st
-from bot.text import parse
+from bot.json import parse
 import asyncio
 import requests
 import io
@@ -50,6 +50,8 @@ async def user_profile(msg: types.Message):
     else:
         user_id = msg.from_user.id
         bot_db = await sqlite.get_id(user_id)
+        display_name = await sqlite.get_display_name(user_id)
+        print(display_name)
         result = await mysql.get_user_profile(bot_db[0])
         date = result[1].strftime("%d.%m.%Y")
         await msg.answer(f"<em>Тариф:</em><b> {result[0]} </b>"
@@ -70,33 +72,48 @@ async def check_inn(msg: types.Message):
 
 
 async def check_result(msg: types.Message, state: FSMContext):
+    test = msg.text
+    if test.isdigit():
+        try:
+            async with state.proxy() as data:
+                data['inn'] = msg.text
+            await msg.delete()
+            await msg.answer(text='Идет сбор данных, ожидайте...')
+            await asyncio.sleep(4)
+            await msg.answer(text="Чтобы получить полную информацию, вы "
+                                  "можете скачать pdf-файл, который будет доступен вместе с ответом")
+            inn = data.get('inn')
+            info = parse.json_parse(inn)
+            if info[9]:
+                json_data = info[9]
+            else:
+                json_data = info[6]
+            user_id = msg.from_user.id
+            u_id = await sqlite.get_id(user_id)
+            text = parse.check_text(info)
+            print(text)
+            if json_data != {'message': 'Компания / ИП не найдены в ЕГРЮЛ / ЕГРИП (2)'}:
+                await mysql.update_log(user_id=u_id, data=inn, json_data=json_data)
+            await msg.answer(text=text, reply_markup=inline.get_pdf_file())
+        except:
+            await msg.answer(text='По введенным данным нет информации, попробуйте ввести другие')
+    else:
+        await msg.answer(text='Неверный формат сообщения. Введите ИНН/ОГРН, используя только цифры.')
+
+
+async def create_pdf(call: types.CallbackQuery, state: FSMContext):
     async with state.proxy() as data:
-        data['inn'] = msg.text
-    await msg.delete()
-    await msg.answer(text='Идет сбор данных, ожидайте')
-    await asyncio.sleep(2)
-    await msg.answer(text="Для получения полной информации вы "
-                          "можете скачать pdf-файл, который будет прикреплен к ответу")
-    inn = data.get('inn')
-    info = parse.json_parse(inn)
-    json = info[9]
-    user_id = msg.from_user.id
-    u_id = await sqlite.get_id(user_id)
-    await mysql.update_log(user_id=u_id, data=inn, json=json)
-    await msg.answer(text=parse.check_text(info))
-    await state.finish()
-
-
-async def get_pdf_file(msg: types.Message):
-    url = "https://svoya-proverka.ru/v2/export-pdf.php?ogrn=1027700132195&" \
-          "blocks=[%221%22,%222%22,%224%22,%225%22,%226%22,%229%22]"
-    response = requests.get(url)
-    with open("example.pdf", "wb") as f:
-        file = f.write(response.content)
-    with open("example.pdf", "rb") as f:
-        file_bytes = io.BytesIO(f.read())
-        input_file = types.InputFile(file_bytes, "Подробный отчет.pdf")
-    await msg.bot.send_document(msg.from_user.id, document=input_file)
+        inn = data.get('inn')
+        url = f"https://svoya-proverka.ru/v2/export-pdf.php?ogrn={inn}&" \
+              f"blocks=[%221%22,%222%22,%224%22,%225%22,%226%22,%229%22]"
+        response = requests.get(url)
+        with open("example.pdf", "wb") as f:
+            file = f.write(response.content)
+        with open("example.pdf", "rb") as f:
+            file_bytes = io.BytesIO(f.read())
+            input_file = types.InputFile(file_bytes, f"Полный отчет_{inn}.pdf")
+        await call.bot.send_document(call.from_user.id, document=input_file)
+        await state.finish()
 
 
 def register(dp: Dispatcher):
@@ -105,6 +122,5 @@ def register(dp: Dispatcher):
     dp.register_message_handler(user_login, commands='login', state='*')
     dp.register_message_handler(user_profile, commands='profile', state='*')
     dp.register_message_handler(check_inn, commands='check', state='*')
-    dp.register_message_handler(get_pdf_file, commands='get_pdf', state="*")
     dp.register_message_handler(check_result, state=st.CheckInn.inn)
-
+    dp.register_callback_query_handler(create_pdf, text='get_pdf_file', state=st.CheckInn.inn)
